+++ internal/interpreter/zz_interpreter_input_test.go
@@ -1,6 +1,7 @@
 package interpreter
 
 import (
+	"basics/internal/common"
 	"basics/internal/constants"
 	"basics/internal/input"
 	"basics/internal/lexer"
@@ -115,7 +116,7 @@ func TestINPUT_TableDriven(t *testing.T) {
 
 			i.Run(prog)
 
-			got := out.String()
+			got := common.StripANSI(out.String())
 			testutils.True(
 				t,
 				fmt.Sprintf("\n--- EXPECTED ---\n%q\n--- GOT ---\n%q\n", tc.want, got),