+++ internal/parser/zz_ast_test.go
@@ -151,6 +151,33 @@ func TestStatementsTypes(t *testing.T) {
 }
 
 // ////////////////////////////////////
+// NORMAL / INVERSE
+// ////////////////////////////////////
+func TestNormalStmt_Pos(t *testing.T) {
+	h := &NormalStmt{
+		Line:   12,
+		Column: 3,
+	}
+
+	line, col, tok := h.Pos()
+	testutils.Equal(t, "line", line, 12)
+	testutils.Equal(t, "column", col, 3)
+	testutils.Equal(t, "token", tok, "NORMAL")
+}
+
+func TestInverseStmt_Pos(t *testing.T) {
+	h := &InverseStmt{
+		Line:   12,
+		Column: 3,
+	}
+
+	line, col, tok := h.Pos()
+	testutils.Equal(t, "line", line, 12)
+	testutils.Equal(t, "column", col, 3)
+	testutils.Equal(t, "token", tok, "INVERSE")
+}
+
+// ////////////////////////////////////
 // HOME / HTAB / VTAB
 // ////////////////////////////////////
 func TestHomeStmt_Pos(t *testing.T) {