+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog (https://keepachangelog.com), and this project adheres to Semantic Versioning (https://semver.org).
 
+## [Unreleased] - 2026-01-29
+### Added
+- Add `NORMAL` support in Apple II Basic. Add relevant unit tests.
+- Add `INVERSE` support in Apple II Basic. Add relevant unit tests.
+
 ## [Unreleased] - 2026-01-28
 ### Added
 - Add `GET` support in Apple II Basic. Add relevant unit tests.